calendar combined view is read only

+++ flutter/lib/view/calendar_month_widget.dart
@@ -160,7 +160,11 @@ class CalenderMonth extends StatelessWidget {
               style: Theme.of(context).textTheme.bodyText2,
               textAlign: TextAlign.center,
             );
+
       onTapHandler = () {
+        if (calendarDetailModel.isReadOnly) {
+          return;
+        }
         developer.log("Pressed index $index");
         if (highlighted) {
           calendarRepository.removeHighlightedDay(
